requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
redis>=5.0.0
chromadb>=0.4.24
slowapi>=0.1.8
websockets>=10.0
prometheus-fastapi-instrumentator
requests
tenacity